requirements.txt
fastmcp
chainlit
strands-agents
strands-agents-tools
strands-agents-builder
boto3